Extract `codec` module
