+++ rust/dns-over-tcp/src/lib.rs
@@ -1,3 +1,4 @@
+mod codec;
 mod interface;
 mod server;
 mod stub_device;
